Adding Character Data User Management and Book Management Evaluation Management Functional Modules Adding Multiple Front-end Pages: Home Page and Personal Center Page Addition, My Evaluations Page Addition, Book Evaluation Page Addition.

+++ book_ratings/routes/web.php
@@ -8,7 +8,11 @@ use App\Http\Controllers\HomeController;
 use App\Http\Controllers\NotificationController;
 use App\Http\Controllers\SearchController;
 use App\Http\Controllers\UserController;
+use App\Http\Controllers\Admin\UserController as AdminUserController;
+use App\Http\Controllers\Admin\BookController as AdminBookController;
+use App\Http\Controllers\Admin\ReviewController as AdminReviewController;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\ReviewController;
 
 /*
 |--------------------------------------------------------------------------
@@ -22,35 +26,43 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('/', function () {
-//    return view('welcome');
-    echo "Hello World";
+    return redirect('/home');
 });
 
 Route::any('/register', [RegisterController::class, 'register'])->name('register');
 Route::any('/login', [LoginController::class, 'login'])->name('login');
 
-Route::get('/profile',[UserController::class, 'profile'])->name('profile')->middleware('auth');
+Route::get('/profile', [UserController::class, 'profile'])->name('profile')->middleware('auth');
 Route::post('/logout', [LoginController::class, 'logout'])->name('logout')->middleware('auth');
 
 Route::get('/home', [HomeController::class, 'index'])->name('home');
+/** three api routes for search */
 Route::any('/search', [SearchController::class, 'search'])->name('search.search');
 
 /**books */
-Route::any('/books/create', [BookController::class, 'create'])->name('books.create');
-Route::any('/books/edit/{id}', [BookController::class, 'edit'])->name('books.edit');
-Route::delete('/books/{id}', [BookController::class, 'delete'])->name('books.destroy');
+Route::any('/books/create', [BookController::class, 'create'])->name('books.create')->middleware('role:admin,author');
+Route::any('/books/edit/{id}', [BookController::class, 'edit'])->name('books.edit')->middleware('role:admin,author');
+Route::delete('/books/{id}', [BookController::class, 'delete'])->name('books.destroy')->middleware('role:admin,author');
 Route::GET('/books', [BookController::class, 'index'])->name('books.index');
 Route::get('/books/show/{id}', [BookController::class, 'show'])->name('books.show');
-Route::post('/books/{book_id}/add_review', [BookController::class, 'add_review'])->name('books.add_review');
-/**reviews */
-Route::get('/my_reviews',[UserController::class, 'myReviews'])->name('myReviews')->middleware('auth');
+Route::post('/books/{book_id}/add_review', [BookController::class, 'add_review'])->name('books.add_review')->middleware('auth');
+Route::get('/books/{book_id}/reviews', [BookController::class, 'reviews'])->name('books.get_reviews');
+/**reviews  */
+Route::get('/my_reviews', [UserController::class, 'myReviews'])->name('myReviews')->middleware('auth');
 Route::get('/users/{user}', [UserController::class, 'show'])->name('users.show');
-// Notifications
+Route::any('/reviews/{id}/edit', [ReviewController::class, 'update_review'])->name('reviews.edit')->middleware('auth');
+Route::delete('/reviews/{id}', [ReviewController::class, 'delete_review'])->name('reviews.destroy')->middleware('auth');
+Route::get('/reviews', [ReviewController::class, 'index'])->name('reviews.index');
+// Notifications Routes
 Route::get('/notifications/unread-count', [NotificationController::class, 'unreadCount'])->name('notifications.unread-count')->middleware('auth');
 Route::get('/notifications/top', [NotificationController::class, 'top'])->name('notifications.top')->middleware('auth');
 Route::get('/notifications', [NotificationController::class, 'index'])->name('notifications')->middleware('auth');
-Route::get('/notifications/{id}', [NotificationController::class, 'show'])->name('notifications.show');
+Route::get('/notifications/{id}', [NotificationController::class, 'show'])->name('notifications.show')->middleware('auth');
 
+/**admin routes */
 Route::prefix('/admin')->middleware('role:admin')->group(function () {
     Route::get('/', [AdminController::class, 'index'])->name('admin.index');
+    Route::resource('/users', AdminUserController::class)->names('admin.users');
+    Route::resource('/reviews', AdminReviewController::class)->names('admin.reviews');
+    Route::resource('/books', AdminBookController::class)->names('admin.books');
 });
